requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
pdfplumber==0.11.0
chromadb==0.5.23
langchain-google-genai==4.2.0
sentence-transformers==3.3.1
pydantic==2.10.4
python-multipart==0.0.20
python-dotenv==1.0.1
